Separate XML tools so SDF code can use them.
